+++ branches/devel-64bit/src/colinux/kernel/monitor.h
@@ -27,10 +27,10 @@ struct co_manager_open_desc;
 
 typedef co_rc_t (*co_monitor_service_func_t)(struct co_monitor *cmon,
 					     struct co_monitor_device *device,
-					     unsigned long *params);
+					     uintptr_t *params);
 typedef struct co_monitor_device {
 	co_monitor_service_func_t service;
-	unsigned long state;
+	uintptr_t state;
 	void *data;
 	co_queue_t out_queue;
 	co_queue_t in_queue;
@@ -91,22 +91,22 @@ typedef struct co_monitor {
 	 */
 	vm_ptr_t	    core_vaddr;  /* Where the core sits (the famous C0100000) */
 	vm_ptr_t	    core_end;    /* Where the core ends */
-	unsigned long	    core_pages;  /* number of pages our core takes */
+	uintptr_t	    core_pages;  /* number of pages our core takes */
 	co_symbols_import_t import;      /* Addresse of symbols in the kernel */
 
 	/*
 	 * Pseudo physical RAM
 	 */
-	unsigned long memory_size;      /* The size of Linux's pseudo physical RAM */
-	unsigned long physical_frames;  /* The number of pages in that RAM */
-	unsigned long end_physical;     /* In what virtual address the map of the
+	uintptr_t memory_size;      /* The size of Linux's pseudo physical RAM */
+	uintptr_t physical_frames;  /* The number of pages in that RAM */
+	uintptr_t end_physical;     /* In what virtual address the map of the
 					   pseudo physical RAM ends */
 	co_pfn_t** pp_pfns;
 
 	/*
 	 * Dynamic allocations in the host
 	 */
-	unsigned long blocks_allocated;
+	uintptr_t blocks_allocated;
 
 	/*
 	 * Page global directory
@@ -175,8 +175,8 @@ typedef struct co_monitor {
         /*
 	 * initrd
 	 */
-	unsigned long initrd_address;
-	unsigned long initrd_size;
+	uintptr_t initrd_address;
+	uintptr_t initrd_size;
 
 	/*
 	 * Structures copied directly from the vmlinux file before it
@@ -195,15 +195,15 @@ struct co_manager_open_desc;
 
 extern co_rc_t co_monitor_ioctl(co_monitor_t*		     cmon,
 				co_manager_ioctl_monitor_t*  io_buffer,
-				unsigned long		     in_size,
-				unsigned long		     out_size,
-				unsigned long*		     return_size,
+				uintptr_t		     in_size,
+				uintptr_t		     out_size,
+				uintptr_t*		     return_size,
 				struct co_manager_open_desc* opened_manager);
 
-extern co_rc_t co_monitor_alloc_pages(co_monitor_t *cmon, unsigned long pages, void **address);
-extern co_rc_t co_monitor_free_pages(co_monitor_t *cmon, unsigned long pages, void *address);
+extern co_rc_t co_monitor_alloc_pages(co_monitor_t *cmon, uintptr_t pages, void **address);
+extern co_rc_t co_monitor_free_pages(co_monitor_t *cmon, uintptr_t pages, void *address);
 
-extern co_rc_t co_monitor_malloc(co_monitor_t *cmon, unsigned long bytes, void **ptr);
+extern co_rc_t co_monitor_malloc(co_monitor_t *cmon, uintptr_t bytes, void **ptr);
 extern co_rc_t co_monitor_free(co_monitor_t *cmon, void *ptr);
 
 extern co_rc_t co_monitor_message_from_user(co_monitor_t *monitor, co_message_t *message);
